Fix gallery
Responsive Lightbox would not work with the Google API version of jQuery,
so the version packaged with WordPress was used instead. Because of this,
the js scripts had to be changed to use jQuery in no conflict mode. The
gallery home was also finished, to look nice without having to load many
images. The gallery page also opens the modal for the album, so no
page redirection is needed.